create popup container
